refactor: eliminate tun_backend_common.cc indirection
